fix: Running redteam from cloud (#2627)
