+++ src/redteam/commands/run.ts
@@ -7,6 +7,7 @@ import telemetry from '../../telemetry';
 import { setupEnv } from '../../util';
 import { getConfigFromCloud } from '../../util/cloud';
 import { doRedteamRun } from '../shared';
+import { getUnifiedConfig } from '../sharedFrontend';
 import type { RedteamRunOptions } from '../types';
 import { poisonCommand } from './poison';
 
@@ -52,8 +53,8 @@ export function redteamRunCommand(program: Command) {
 
       let autoShare = false;
       if (opts.config && UUID_REGEX.test(opts.config)) {
-        const config = await getConfigFromCloud(opts.config);
-        opts.liveRedteamConfig = config;
+        const configObj = await getConfigFromCloud(opts.config);
+        opts.liveRedteamConfig = getUnifiedConfig(configObj.config);
         opts.config = undefined;
         autoShare = true;
       }
